+++ tests/test_workflow.py
@@ -15,7 +15,7 @@
 # along with pytest-workflow.  If not, see <https://www.gnu.org/licenses/
 
 """Tests the Workflow class"""
-import subprocess  # nosec. This is just for the Timeout exception.
+import subprocess
 
 import pytest
 
@@ -96,3 +96,12 @@ def test_workflow_name():
 def test_workflow_name_inferred():
     workflow = Workflow("echo moo")
     assert workflow.name == "echo"
+
+
+def test_workflow_matching_exit_code():
+    workflow = Workflow("echo moo")
+    workflow.run()
+    assert workflow.matching_exitcode()
+    workflow2 = Workflow("grep", desired_exit_code=2)
+    workflow2.run()
+    assert workflow2.matching_exitcode()